backend/pyproject.toml
[project]
name = "rides-coordinator-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "apscheduler>=3.11.0",
    "discord>=2.3.2",
    "fastapi>=0.115.0",
    "greenlet>=3.2.4",
    "httpx>=0.28.1",
    "invoke>=2.2.0",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.10",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.3.0",
    "pytz>=2025.2",
    "rapidfuzz>=3.14.3",
    "recurring-ical-events>=3.8.0",
    "sqlalchemy>=2.0.43",
    "tenacity>=9.1.2",
    "uvicorn[standard]>=0.34.0",
]

[dependency-groups]
dev = [
    "invoke>=2.2.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.13.0",
]
